refactor: directory names
