+++ src/components/header/ScrollProgressBar.tsx
