+++ src/components/post/CardImage.tsx
