+++ src/types/announcement.ts
